backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
PyPDF2==3.0.1
pdfplumber==0.10.3
alembic==1.12.1
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
pydantic==2.5.0
pydantic-settings==2.1.0
python-dateutil==2.8.2
redis==5.0.1
apscheduler==3.10.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
rapidfuzz==3.5.2
python-multipart==0.0.6
pyyaml==6.0.1